fastlane configuration module require improvements & fixes
