+++ fastlane_core/lib/fastlane_core/configuration/commander_generator.rb
@@ -1,5 +1,7 @@
 require 'commander'
-require 'fastlane/version'
+
+require_relative '../module'
+require_relative '../ui/ui'
 
 module FastlaneCore
   class CommanderGenerator
@@ -30,7 +32,7 @@ module FastlaneCore
 
         # Boolean is a fastlane thing, it's either TrueClass, or FalseClass, but we won't know
         # that until runtime, so nil is the best we get
-        if type == Boolean
+        if type == Fastlane::Boolean
           type = nil
         end
