+++ fastlane_core/lib/fastlane_core/ui/errors/fastlane_error.rb
@@ -1,3 +1,5 @@
+require_relative 'fastlane_exception'
+
 module FastlaneCore
   class Interface
     class FastlaneError < FastlaneException
